use text files from pre-existing meta.json

+++ train_mixed_tokenizer.py
@@ -1,6 +1,11 @@
+"""
+Train tokenizer on a mixture of data categories. The categories and their weights are sampled from the unit simplex.
+"""
+
 import os
 from pathlib import Path
 import time
+import re
 import json
 import random
 from collections import defaultdict, Counter
@@ -34,39 +39,39 @@ def sample_from_unit_simplex(n, M=10000):
     help='Whether to use SentencePiece library for training. Default is tokenizers.'
 )
 @click.option(
-    '--num_languages',
+    '--num_categories',
     type=int,
-    default=40,
+    default=10,
     help='Number of languages to include in the mixed tokenizer training set. Weights will be sampled.'
 )
 @click.option(
     '--total_bytes',
     type=int,
     default=10**10,
-    help='The maximum number of characters to use for tokenizer training.'
+    help='The maximum number of bytes to use for tokenizer training.'
 )
 @click.option(
     '--corpus_dir',
     type=str,
     default='/gscratch/scrubbed/alisaliu/oscar-corpus/processed',
-    help='Directory containing language subdirectories with text files, to use for training the tokenizer.'
+    help='Directory containing subdirectories for each category, to use for training the tokenizer.'
 )
 @click.option(
     '--use_wiki_languages',
     type=bool,
     default=False,
-    help="Whether to use only the set of languages available in RedPajama's wikipedia split."
+    help="Whether to use only the set of languages available in RedPajama's wikipedia split. This is used for the distribution shift experiment."
 )
 @click.option(
     '--size_threshold',
     type=int,
     default=None,
-    help='If specified, we will only use languages with byte counts above this threshold.'
+    help='If specified, we will only use languages with byte counts above this threshold. This is used for the data size ablation.'
 )
 def main(
     output_dir: str,
     use_spm: bool,
-    num_languages: int,
+    num_categories: int,
     total_bytes: int,
     corpus_dir: str,
     use_wiki_languages: bool,
@@ -77,7 +82,7 @@ def main(
     ensure_dir(output_dir)
     print(f'We are training a tokenizer for {output_dir}', flush=True)
 
-    # If meta.json exists, use the languages and weights from there.
+    # If meta.json exists, use the mixture and files from there.
     if os.path.exists(output_dir / 'meta.json'):
         print('Output directory contains meta.json, so we will use the languages and weights from there.', flush=True)
         meta = json.load(open(output_dir / 'meta.json'))
@@ -85,34 +90,41 @@ def main(
         languages = byte_counts.keys()
         weights = [v / sum(byte_counts.values()) for v in byte_counts.values()]
         train_files = meta['train_files']
-        text_files_ok = all([os.path.exists(f) for f in train_files])
-        if text_files_ok:
-            text_files = []
-            for item, count in train_files.items():
-                text_files.extend([item] * count)
+        text_files = []
+        for file, count in train_files.items():
+            # file = file.replace('/gscratch/scrubbed/alisaliu/redpajama/github', '/gscratch/xlab/alisaliu/redpajama/github')
+            if not os.path.exists(file):
+                assert 'truncated' in file, f'{file} not found'
+                untruncated_file = re.sub(r'_truncated_\d+', '', file)
+                wanted_filesize = int(re.search(r'_truncated_(\d+)', file).group(1))
+                os.system(f'cp {untruncated_file} {file}')
+                truncate_file(file, wanted_filesize)
+            text_files.extend([file] * count)
 
     # Else, sample languages and weights.
     else:
-        text_files_ok = False
         if use_wiki_languages:
             print('Using only languages available in RedPajama wikipedia split.', flush=True)
-            languages = random.sample(os.listdir('/gscratch/scrubbed/alisaliu/redpajama/wikipedia'), num_languages)
+            languages = random.sample(os.listdir('/gscratch/xlab/alisaliu/redpajama/wikipedia'), num_categories)
         elif size_threshold:
-            from constants import LANG_SIZES
-            all_languages = [lang for lang, size in LANG_SIZES.items() if size > size_threshold]
-            languages = random.sample(all_languages, num_languages)
+            from constants import keyword_to_mixture_type
+            for keyword in keyword_to_mixture_type:
+                if keyword in str(output_dir):
+                    CATEGORY_SIZES = keyword_to_mixture_type[keyword]
+
+            all_languages = [lang for lang, size in CATEGORY_SIZES.items() if size > size_threshold]
+            languages = random.sample(all_languages, num_categories)
             print(f'Using the {len(all_languages)} languages with at least {size_threshold} bytes of text.', flush=True)
         else:
-            languages = random.sample(os.listdir(corpus_dir), num_languages)
+            languages = random.sample(os.listdir(corpus_dir), num_categories)
 
         weights = sample_from_unit_simplex(len(languages))
         print(f'Intended language distribution: {({l: w for l, w in zip(languages, weights)})}', flush=True)
 
-    if not text_files_ok:
         tqdm_bar = tqdm(total=total_bytes, desc='Loading text data')
         text_files, byte_counts = defaultdict(list), defaultdict(int)
 
-        # get text data for each language, duplicating as necessary until byte_counts for each language is met
+        # Get text data for each language, duplicating as necessary until byte_counts for each language is met.
         for lang_code, weight in zip(languages, weights):
             language_files = [f for f in os.listdir(corpus_dir / lang_code) if f.endswith('.txt') and 'truncated' not in f]
             random.shuffle(language_files)
@@ -135,7 +147,7 @@ def main(
                     tqdm_bar.update(wanted_filesize)
                 counter += 1
 
-        # Write metadata
+        # Write metadata.
         text_files = [f for lang_files in text_files.values() for f in lang_files]
         with open(output_dir / 'meta.json', 'w') as fo:
             meta = {}
@@ -154,6 +166,7 @@ def main(
         print('Training with HF tokenizers...')
         tokenizer = train_tokenizer_or_dump_frequencies(text_files)
         tokenizer.model.save(str(output_dir))
+        tokenizer.save(str(output_dir / 'tokenizer.json'))
     else:
         from utils import train_tokenizer_spm
         print('Training with SentencePiece...')
@@ -169,7 +182,7 @@ def main(
 
     # Delete files that were constructed just for this
     for f in text_files:
-        if 'truncated' in f:
+        if os.path.exists(f) and 'truncated' in f:
             os.remove(f)
